+++ src/InvalidArgumentException.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Skautis;
+
+
+/**
+ * @author Hána František <sinacek@gmail.com>
+ */
+class InvalidArgumentException extends \InvalidArgumentException implements Exception
+{
+
+}