+++ src/Wsdl/AuthenticationException.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Skautis\Wsdl;
+
+
+/**
+ * Neplatné přihlášení spojené se skautISem.
+ *
+ * @author Hána František <sinacek@gmail.com>
+ */
+class AuthenticationException extends WsdlException
+{
+
+}