+++ src/Wsdl/PermissionException.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Skautis\Wsdl;
+
+
+/**
+ * Nepovolený přístup ze strany SkautISu.
+ *
+ * @author Hána František <sinacek@gmail.com>
+ */
+class PermissionException extends WsdlException
+{
+
+}