+++ src/Wsdl/WsdlException.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Skautis\Wsdl;
+
+use Skautis;
+
+
+/**
+ * Obecná chyba při komunikaci s webovými službami.
+ *
+ * @author Hána František <sinacek@gmail.com>
+ */
+class WsdlException extends \Exception implements Skautis\Exception
+{
+
+}